Add Repository layer with JDBC

+++ src/main/kotlin/org/example/db_project_2024/model/User.kt
@@ -1,7 +1,7 @@
 package org.example.db_project_2024.model
 
-data class User (
-    val id: Int,
+data class User(
+    val id: Long,
     val firstName: String,
     val lastName: String,
     val password: String,
